fix: resolve bug module rematch won't work if player 2 win
