+++ lib/menu.c
@@ -1,6 +1,5 @@
 #include "menu.h"
 #include "game.c"
-#include "score.c"
 
 /**
  * Procedure for setting game mode whether it's player vs player OR player vs computer.
@@ -28,7 +27,8 @@ void selectGameMode(bool err, int *opsiMode)
   {
     selectGameMode(1, opsiMode);
   }
-  else if (*opsiMode == 0){
+  else if (*opsiMode == 0)
+  {
     playMenu();
   }
 }
@@ -87,46 +87,6 @@ void selectBoard(bool err, int *opsiBoard)
   if (*opsiBoard != 1 && *opsiBoard != 2 && *opsiBoard != 3)
   {
     selectBoard(1, opsiBoard);
-  }
-}
-
-/**
- * Procedure for rematch the game.
- */
-void rematch(bool err, int mode, int bot, int board, char *player1, char *player2)
-{
-
-  int input;
-
-  // Out the error msg
-  if (err)
-  {
-    printf("\033[1;31m[ERR]\t\033[1;0mInvalid input. Try again.");
-  }
-
-  printf("\n\n1. Rematch");
-  printf("\n2. Go back to main menu");
-  printf("\n3. Exit game");
-  printf("\n\nPlease enter a number between 1-3: ");
-  scanf("%d", &input);
-  getchar(); // clear input; resolve the infinity loop
-
-  // conditional for user input
-  switch (input)
-  {
-  case 1:
-    tictactoe(mode, bot, board, player1, player2);
-    break;
-  case 2:
-    menu(0);
-    break;
-  case 3:
-    exit(0);
-    break;
-  default:
-    // if user not type 1 OR 2 OR 3; then repreat this procedure with error msg
-    rematch(1, mode, bot, board, player1, player2);
-    break;
   }
 }
 
@@ -204,8 +164,6 @@ void playMenu()
     playername2 = "";
   }
   tictactoe(mode, bot, board, playername1, playername2);
-
-  rematch(0, mode, bot, board, playername1, playername2);
 }
 
 /**
